Add files via upload
New York City is often celebrated as a melting pot, a vibrant mosaic of cultures and communities that enrich the urban landscape. However, beneath this colorful exterior lies a pressing issue: the housing challenges faced by immigrant populations. More critical to the cause of the city government in its bid to provide affordable housing in the Yes City project, it has become necessary to analyze the clusters of immigrant neighborhoods in NYC, their unique characteristics, and the significant hurdles immigrants encounter in securing affordable housing to inform policy direction of a city with almost half of its population being immigrants.